+++ src/main/java/easeplan/netease/sales/domain/CartItem.java
@@ -3,6 +3,8 @@ package easeplan.netease.sales.domain;
 import lombok.Data;
 
 /**
+ * 购物车内容
+ *
  * @author huangzw
  * @version 1.0
  * @since <pre>2018/3/7</pre>
@@ -11,6 +13,6 @@ import lombok.Data;
 public class CartItem {
     private int id;
     private String title;
-    private int price;
+    private long price;
     private int amount;
 }
