+++ src/main/java/easeplan/netease/sales/domain/PurchasedItem.java
@@ -3,6 +3,8 @@ package easeplan.netease.sales.domain;
 import lombok.Data;
 
 /**
+ * 内容购买记录
+ *
  * @author huangzw
  * @version 1.0
  * @since <pre>2018/3/7</pre>
@@ -13,6 +15,6 @@ public class PurchasedItem {
     private String image;
     private String title;
     private long purchaseDate;
-    private int purchasePrice;
+    private long purchasePrice;
     private int purchaseAmount;
 }
